+++ packages/stream_chat_persistence/lib/src/dao/channel_query_dao.dart
@@ -112,8 +112,9 @@ class ChannelQueryDao extends DatabaseAccessor<MoorChatDatabase>
 
     cachedChannels.sort(chainedComparator);
 
-    if (paginationParams?.offset != null && cachedChannels.isNotEmpty) {
-      cachedChannels.removeRange(0, paginationParams!.offset);
+    final offset = paginationParams?.offset;
+    if (offset != null && offset > 0 && cachedChannels.isNotEmpty) {
+      cachedChannels.removeRange(0, offset);
     }
 
     if (paginationParams?.limit != null) {