+++ packages/stream_chat/CHANGELOG.md
@@ -13,6 +13,8 @@
 ✅ Added
 
 - Added `Filter.contains` and `Filter.empty`
+- Added support for `next`, `previous` value pagination in `client.search`
+  , [read more.](https://getstream.io/chat/docs/other-rest/search/#pagination)
 
 ## 2.2.1